Combine entirely
